pyproject.toml
[tool.poetry]
name = "weatherdata"
version = "0.1.0"
description = "Import weather from text file and show it in the page"
authors = ["Meena Bharani <ns.BharaniMeena@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.6"
djangorestframework = "^3.14.0"
markdown = "^3.4.1"
django-filter = "^22.1"
psycopg2 = "^2.9.5"
psycopg2-binary = "^2.9.5"
django-cors-headers = "^3.13.0"
pandas = "1.4.0"
python-dotenv = "^0.21.1"
pip = "^23.0"
sqlalchemy = "^2.0.2"
django-environ = "^0.9.0"
django-rest-swagger = "^2.2.0"
drf-yasg = "^1.21.4"
pylint = "^2.16.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
